requirements.txt
# FastAPI + Uvicorn
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Pydantic v2 + settings
pydantic==2.9.2
pydantic-settings==2.5.2

# JWT
PyJWT==2.9.0

# Password hashing
bcrypt==4.2.0

# PostgreSQL
psycopg2-binary==2.9.9

# Redis
redis==5.1.0

# DSPy
dspy-ai==2.5.7

# Anthropic (Claude API — DSPy backbone + RLHF refinement)
anthropic==0.36.2

# SQLAlchemy (RLHF SQLite ORM)
SQLAlchemy==2.0.35

# HTTP client (Cube.js calls)
requests==2.32.3
httpx==0.27.2

# YAML parsing (catalog, semantic layer)
PyYAML==6.0.2

# OpenTelemetry tracing
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
openinference-instrumentation-anthropic==0.1.17

# Scheduling (RLHF refinement cycle)
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.1
colorlog==6.8.2
